Don't call the module onload callback if the early onload callback fails.
When goog.module.ModuleManager loads a module, it registers a callback and an
errback with the module info object, which invoke the callback and errback
methods (respectively) of the deferred object that tracks the module load.

With the current behavior of goog.module.ModuleInfo, an error in a module
initialization callback causes both the callback and the errback of the
deferred to be invoked, which raises a goog.async.Deferred.AlreadyCalledError
during the errback call.  (The module info object then catches this error and
handles it by raising a "Module errback failures" error inside a setTimeout.)

R=pupius
DELTA=1  (0 added, 0 deleted, 1 changed)


Revision created by MOE tool push_codebase.
MOE_MIGRATION=3026


git-svn-id: http://closure-library.googlecode.com/svn/trunk@1212 0b95b8e8-c90f-11de-9d4f-f947ee5921c8