Ingest http://codereview.appspot.com/5229042/ from Rich Dougherty
"When testing with goog.testing.net.XhrIo we can use the getLastXxx() methods to
find the parameters that were used in the last send() call. There are methods
for accessing all but one of send()'s parameters. Unfortunately access to the
'method' parameter is missing. I'd like to be able to test that my code is using
the correct method (e.g. 'PUT') so I've written a patch that adds
getLastMethod() to access that information."

Fixes issue 384.

R=nicksantos
DELTA=44  (33 added, 2 deleted, 9 changed)


Revision created by MOE tool push_codebase.
MOE_MIGRATION=3444


git-svn-id: http://closure-library.googlecode.com/svn/trunk@1328 0b95b8e8-c90f-11de-9d4f-f947ee5921c8