Give the developer a clearer signal that a test has failed.
Recently I checked in a broken test because when I ran the
test locally in my browser via testrunner, the only failure
was off the bottom of the page and I missed the one red line
at the top of the page in the sea of green.  So I propose that
we highlight the page on failure to alert the developer.

R=nnaze
DELTA=12  (11 added, 0 deleted, 1 changed)


Revision created by MOE tool push_codebase.
MOE_MIGRATION=5079


git-svn-id: http://closure-library.googlecode.com/svn/trunk@2009 0b95b8e8-c90f-11de-9d4f-f947ee5921c8