Automated g4 rollback
*** Reason for rollback ***

After the CL, goog.Disposable is not a no-op, and cannot be compiled
out. This is unacceptable.

*** Original change description ***

Disposable monitoring mode: Enable starting the monitoring mode on the fly.

Some Disposables might be created before the monitoring mode was started,
don't throw an error when they are disposed.

To catch the "base class constructor not called" error in the monitoring mode,
make disposed_ undefined unless the base class constructor was called, and
check

... description truncated by g4 rollback ...

R=ssaviano


Revision created by MOE tool push_codebase.
MOE_MIGRATION=5154


git-svn-id: http://closure-library.googlecode.com/svn/trunk@2037 0b95b8e8-c90f-11de-9d4f-f947ee5921c8